+++ apps/meteor/app/reactions/client/index.ts
@@ -1,2 +1,1 @@
-import './useReactionMessageAction';
 import './methods/setReaction';
